Apps/Claims/requirements.txt
SQLAlchemy==2.0.42
python-dotenv==1.1.1
pydantic==2.11.7
pytesseract==0.3.13
pdf2image==1.17.0
PyPDF2==3.0.1
pandas==2.3.1
langchain-core==0.3.72
langchain==0.3.27
RapidFuzz==3.13.0
pdfplumber==0.11.6
torch==2.7.1
sentence-transformers==5.0.0
openpyxl==3.1.5
transformers
langchain-ollama==0.3.6
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
asyncpg==0.30.0
